+++ packages/algob/test/internal/core/config/config-validation.ts
@@ -312,6 +312,18 @@ describe("Config validation", function () {
 						assert.isEmpty(errors.toString());
 					});
 
+					it("Token can be null", function () {
+						const errors = getValidationErrors({
+							networks: {
+								custom: {
+									host: "http://localhost",
+									token: null,
+								},
+							},
+						});
+						assert.isEmpty(errors.toString());
+					});
+
 					it("should pass with passing X-Algo-API-Token", function () {
 						const errors = getValidationErrors({
 							networks: {