+++ CHANGELOG.md
@@ -83,12 +83,14 @@ Added:
 - `printGlobalStateSCC` renamed to `printGlobalStateApp`.
 
 - The ` PyASCCache` has been merged to `ASCCache` and is not used anymore.
+- Rename the executeTransaction to executeTx
 
 ### Bug fixes
 
 - Return error when closeRemainderTo and fromAccountAddr is the same.
 - When close account should remove auth/spend address. Fixed in [#575](https://github.com/scale-it/algo-builder/pull/575).
 - Approval program and clear propram should throw error if they are mismatch version. Fixed in [620](https://github.com/scale-it/algo-builder/pull/620)
+- Allow token to be empty
 ### Infrastructure
 
 - Updated `setup-master-account` and `sandbox-setup-master-account` commands to run multiple times.